requirements.txt
beautifulsoup4==4.10.0
boto3==1.20.24
Django==3.2.5
django-cors-headers==3.10.1
django-mysql==4.3.0
djangorestframework==3.13.1
Jinja2==3.0.3
knox==0.1.14
mysqlclient==2.1.0
pandas==1.3.5
selenium==4.1.0
transformers==4.5.1
pillow==8.4.0
scikit-learn
